Harmonize Server controller

+++ src/sdk-reference/cpp/1/server/now/snippets/now.cpp
@@ -1,5 +1,6 @@
 try {
-  int64_t timestamp = kuzzle->server->now();
+  long long timestamp = kuzzle->server->now();
+
   std::cout << "Epoch-millis timestamp: " << timestamp << std::endl;
 } catch (kuzzleio::KuzzleException& e) {
   std::cerr << e.what() << std::endl;
